Refactor TurboQuant encoding API and tests
- Use ExecutionCtx in TurboQuant compress path and import ExecutionCtx
- Extend dtype imports with Nullability and PType to support extension
  types
- Wire in extension utilities: extension_element_ptype and
  extension_list_size for vector extensions
- Remove dimension and bit_width from slice/take compute calls to rely
  on metadata
- Update TurboQuant mod docs to mention VortexSessionExecute
- Change scheme.compress to use the provided compressor argument (not
  _compressor)
- Add an extensive TurboQuant test suite (roundtrip, MSE bounds, edge
  cases, f64 input, serde roundtrip, and dtype checks)
- Align vtable imports to new metadata handling (remove unused
  DeserializeMetadata/SerializeMetadata references)
