+++ vortex-tensor/src/encodings/turboquant/compute/take.rs
@@ -27,8 +27,6 @@ impl TakeExecute for TurboQuant {
             taken_norms,
             array.centroids().clone(),
             array.rotation_signs().clone(),
-            array.dimension,
-            array.bit_width,
         )?;
 
         Ok(Some(result.into_array()))
